+++ R/data.R
@@ -1,0 +1,13 @@
+#' Example prioritised targets
+#'
+#' @description
+#' Example output from the function \link[MultiEWCE]{prioritise_targets}.
+#' Used in examples to reduce run time.
+#' @source
+#' \code{
+#' example_targets <- prioritise_targets()
+#' usethis::use_data(example_targets, overwrite = TRUE)
+#' }
+#' @format ontology_index
+#' @usage data("example_targets")
+"example_targets"