+++ R/plot_visnetwork.R
@@ -1,31 +1,39 @@
 plot_visnetwork <- function(g,
-                            save_path = tempfile(
-                              fileext =
-                                "_prioritise_targets_network.html"
-                            ),
-                            layout = "layout_with_kk",
-                            solver = "forceAtlas2Based",
-                            physics = FALSE,
-                            forceAtlas2Based = list(
-                              avoidOverlap=.5,
-                              gravitationalConstant=-50),
-                            scaling=list(label=
-                                           list(drawThreshold=0)
-                            ),
-                            smooth=list(enabled=TRUE,
-                                        type="cubicBezier",
-                                        roundness=.5),
-                            add_visExport = FALSE,
-                            height = 1000,
-                            width = 1300,
-                            randomSeed = 2023,
-                            verbose=TRUE){
+                            save_path,
+                            layout,
+                            solver,
+                            physics,
+                            forceAtlas2Based,
+                            scaling,
+                            smooth,
+                            add_visExport,
+                            degree,
+                            height,
+                            width,
+                            randomSeed,
+                            main,
+                            submain,
+                            verbose){
   requireNamespace("visNetwork")
+  requireNamespace("dplyr")
+
+  . <- NULL;
 
   messager("Creating plot.",v=verbose)
-  visnet <- visNetwork::visIgraph(g,
-                                  randomSeed = randomSeed) |>
+  visnet <- visNetwork::toVisNetworkData(g) %>%
+    {
+      do.call(visNetwork::visNetwork,
+              c(., list(main = main,
+                        submain = submain,
+                        background = "transparent")
+                )
+              )
+    } |>
+    # visNetwork::visIgraph(g,
+    #                               randomSeed = randomSeed) |>
     visNetwork::visIgraphLayout(layout = layout,
+                                type = "full",
+                                randomSeed = randomSeed,
                                 physics = physics) |>
     visNetwork::visPhysics(solver=solver,
                            forceAtlas2Based=forceAtlas2Based,
@@ -33,12 +41,16 @@ plot_visnetwork <- function(g,
     visNetwork::visNodes(font = list(color="#F0FFFF",
                                      strokeWidth=2,
                                      strokeColor="rgba(0,0,0,1)"
-    ),
+                                     ),
     shadow = list(enabled=TRUE,
                   size = 10),
     opacity = 0.75,
-    color = list(hover=list(background="rgba(0,0,0,.5)")
-    ),
+    borderWidth=3,
+    borderWidthSelected=6,
+    color = list(hover = list(background="rgba(0,0,0,.5)"),
+                 highlight = list(background="#00FFFFCF",
+                                  border="#00FFFFCF")
+                 ),
     scaling = scaling
     ) |>
     visNetwork::visEdges(shadow = list(enabled=FALSE),
@@ -50,11 +62,13 @@ plot_visnetwork <- function(g,
     #                       color="green")
     # visNetwork::visClusteringByGroup(groups = igraph::vertex_attr(g,"group"))
     visNetwork::visInteraction(hover = TRUE) |>
-    visNetwork::visOptions(height=height,
+    visNetwork::visOptions(height = height,
                            width = width,
+                           selectedBy = "node",
                            highlightNearest = list(enabled=TRUE,
-                                                   degree=2))
-  if(add_visExport){
+                                                   degree=degree))
+
+  if(isTRUE(add_visExport)){
     visnet <-  visnet |> visNetwork::visExport(type = "pdf")
   }
   #### Save network ####