+++ tests/testthat/test-correlation_heatmap.R
@@ -1,0 +1,7 @@
+test_that("correlation_heatmap works", {
+
+  top_targets <- example_targets$top_targets
+  hm <- correlation_heatmap(top_targets = top_targets)
+
+  testthat::expect_true(methods::is(hm,"Heatmap"))
+})